+++ tests/testthat/test-basics.R
@@ -347,6 +347,17 @@ test_that("recipe() errors if `data` is missing", {
   expect_snapshot(error = TRUE, recipe(mpg ~ .))
 })
 
+test_that("NAs aren't dropped in strings2factor() (#1291)", {
+  ex_data <- tibble(
+    x = factor(c("a", NA, "c"), exclude = NULL)
+  )
+
+  rec_res <- recipe(~., data = ex_data) %>%
+    prep() %>%
+    bake(new_data = NULL)
+
+  expect_identical(rec_res, ex_data)
+})
 
 test_that("recipe() can handle very long formulas (#1283)", {
   df <- matrix(1:10000, ncol = 10000)